Renamed algorithms package to token
