+++ oauth/marshal.go
@@ -3,7 +3,7 @@ package oauth
 import (
 	"fmt"
 	"github.com/golang-jwt/jwt/v5"
-	"github.com/stevezaluk/credstack-lib/oauth/algorithm"
+	"github.com/stevezaluk/credstack-lib/oauth/token"
 	"github.com/stevezaluk/credstack-lib/proto/response"
 )
 
@@ -15,12 +15,12 @@ TODO: Need support for id tokens and refresh tokens here
 TODO: Expires in is not rendering properly, showing expiration instead of token lifetime
 TODO: This function feels kind of clunky...
 */
-func MarshalTokenResponse(token *jwt.Token, signedString string) (*response.TokenResponse, error) {
-	expirationDate, err := token.Claims.GetExpirationTime()
+func MarshalTokenResponse(tok *jwt.Token, signedString string) (*response.TokenResponse, error) {
+	expirationDate, err := tok.Claims.GetExpirationTime()
 	if err != nil {
 		// wrapping this error with ErrFailedToSignToken is not ideal as it can lead to some confusion on
 		// how this function failed but... oh well!
-		return nil, fmt.Errorf("%w (%v)", algorithm.ErrFailedToSignToken, err)
+		return nil, fmt.Errorf("%w (%v)", token.ErrFailedToSignToken, err)
 	}
 	/*
 		After we actually sign our token, we can quickly convert it back into a response.TokenResponse structure so that
