+++ oauth/token/rs256.go
@@ -1,4 +1,4 @@
-package algorithm
+package token
 
 import (
 	"fmt"
